+++ client/src/util/patterns.ts
@@ -5,7 +5,10 @@ export default {
 		ENTITY_GLOBAL: /^&mvt[a-z]?:global:(.+?);$/i,
 
 		VARIABLE_LOCAL: /^l\.settings\:(.+?)$/i,
-		VARIABLE_GLOBAL: /^g\.(.+?)/i
+		VARIABLE_GLOBAL: /^g\.(.+?)/i,
+
+		FOREACH_TAG_OPEN: /<mvt:foreach/gi,
+		FOREACH_TAG_CLOSE: /<\/mvt:foreach/gi
 
 	},
 	MV: {